🧹 Refactor InfrastructureOptimizerAgent into smaller modules
