+++ packages/api/src/services/ai-agents/infrastructure-optimizer.ts
@@ -11,24 +11,13 @@
 import { BaseAgent } from "./orchestrator";
 import { logError, logInfo } from "../../utils/logger";
 import { prisma } from "../../infrastructure/db/prisma";
+import { OptimizationOpportunity } from "./infrastructure-optimizer/types";
+import { analyzeQueries } from "./infrastructure-optimizer/query-analyzer";
+import { analyzeCosts } from "./infrastructure-optimizer/cost-analyzer";
+import { analyzePerformance } from "./infrastructure-optimizer/performance-analyzer";
+import { analyzeCapacity } from "./infrastructure-optimizer/capacity-analyzer";
 
-export interface OptimizationOpportunity {
-  id: string;
-  type: "query" | "cost" | "performance" | "capacity";
-  description: string;
-  currentState: Record<string, unknown>;
-  proposedChange: Record<string, unknown>;
-  expectedImpact: {
-    costSavings?: number;
-    performanceImprovement?: number;
-    errorRateReduction?: number;
-    memoryReduction?: number;
-    throughputIncrease?: number;
-    loadReduction?: number;
-    riskLevel: "low" | "medium" | "high";
-  };
-  recommendedAction: "auto-apply" | "human-review";
-}
+export { OptimizationOpportunity };
 
 export class InfrastructureOptimizerAgent extends BaseAgent {
   id = "infrastructure-optimizer";
@@ -100,19 +89,19 @@ export class InfrastructureOptimizerAgent extends BaseAgent {
     const opportunities: OptimizationOpportunity[] = [];
 
     // Analyze query performance
-    const slowQueries = await this.findSlowQueries();
+    const slowQueries = await analyzeQueries();
     opportunities.push(...slowQueries);
 
     // Analyze costs
-    const costOpportunities = await this.findCostOptimizations();
+    const costOpportunities = await analyzeCosts();
     opportunities.push(...costOpportunities);
 
     // Analyze performance
-    const performanceOpportunities = await this.findPerformanceIssues();
+    const performanceOpportunities = await analyzePerformance();
     opportunities.push(...performanceOpportunities);
 
     // Analyze capacity
-    const capacityOpportunities = await this.findCapacityIssues();
+    const capacityOpportunities = await analyzeCapacity();
     opportunities.push(...capacityOpportunities);
 
     this.optimizationHistory = opportunities;
@@ -128,363 +117,6 @@ export class InfrastructureOptimizerAgent extends BaseAgent {
           logError(`Failed to apply optimization ${opportunity.id}`, error);
         });
       }
-    }
-
-    return opportunities;
-  }
-
-  /**
-   * Find slow queries by querying pg_stat_statements (if available)
-   */
-  private async findSlowQueries(): Promise<OptimizationOpportunity[]> {
-    const opportunities: OptimizationOpportunity[] = [];
-
-    try {
-      const slowQueries = (await prisma.$queryRaw`
-        SELECT query, mean_exec_time, calls, rows
-        FROM pg_stat_statements
-        WHERE mean_exec_time > 100
-        ORDER BY mean_exec_time DESC
-        LIMIT 10
-      `) as any[];
-
-      for (let i = 0; i < slowQueries.length; i++) {
-        const sq = slowQueries[i];
-        if (!sq) continue;
-        if (
-          sq.query.includes("CREATE INDEX") ||
-          sq.query.startsWith("COMMIT") ||
-          sq.query.startsWith("BEGIN")
-        ) {
-          continue;
-        }
-
-        opportunities.push({
-          id: `opt_query_${i + 1}`,
-          type: "query",
-          description: `Slow query detected: ${sq.query.substring(0, 100)}...`,
-          currentState: {
-            query: sq.query,
-            avgDuration: sq.mean_exec_time,
-            callCount: sq.calls,
-          },
-          proposedChange: {
-            recommendation: "Consider adding indexes or optimizing query structure",
-          },
-          expectedImpact: {
-            performanceImprovement: 50,
-            riskLevel: "low" as const,
-          },
-          recommendedAction: "human-review" as const,
-        });
-      }
-    } catch {
-      // pg_stat_statements not available, fall back to query log analysis
-      logInfo("pg_stat_statements not available, using query log analysis");
-
-      const recentJobs = await prisma.reconJob.findMany({
-        where: { createdAt: { gte: new Date(Date.now() - 24 * 60 * 60 * 1000) } },
-        select: { id: true, executionTime: true },
-        orderBy: { executionTime: "desc" },
-        take: 20,
-      });
-
-      for (let i = 0; i < recentJobs.length; i++) {
-        const job = recentJobs[i];
-        if (job.executionTime && job.executionTime > 60000) {
-          opportunities.push({
-            id: `opt_slow_job_${i + 1}`,
-            type: "query",
-            description: `Slow reconciliation job: ${job.id}`,
-            currentState: {
-              jobId: job.id,
-              executionTimeMs: job.executionTime,
-            },
-            proposedChange: {
-              recommendation: "Review job configuration and data volume",
-            },
-            expectedImpact: {
-              performanceImprovement: 30,
-              riskLevel: "low" as const,
-            },
-            recommendedAction: "human-review" as const,
-          });
-        }
-      }
-    }
-
-    return opportunities;
-  }
-
-  /**
-   * Find cost optimization opportunities
-   */
-  private async findCostOptimizations(): Promise<OptimizationOpportunity[]> {
-    const opportunities: OptimizationOpportunity[] = [];
-
-    // 1. Analyze AI usage patterns
-    const thirtyDaysAgo = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
-    const aiCalls = await prisma.usageEvent.findMany({
-      where: {
-        eventType: "ai_request",
-        timestamp: { gte: thirtyDaysAgo },
-      },
-      select: { metadata: true },
-    });
-
-    // Group by model and find optimization opportunities
-    const modelUsage: Record<string, { calls: number; tokens: number; cost: number }> = {};
-    for (const call of aiCalls) {
-      const metadata = (call.metadata as any) || {};
-      const model = metadata.model || "unknown";
-      const cost = metadata.cost || 0;
-      const tokens = metadata.tokens || 0;
-
-      if (!modelUsage[model]) {
-        modelUsage[model] = { calls: 0, tokens: 0, cost: 0 };
-      }
-      const usage = modelUsage[model]!;
-      usage.calls++;
-      usage.tokens += tokens;
-      usage.cost += cost;
-    }
-
-    // Check for expensive model usage that could be downgraded
-    const gpt4Usage = modelUsage["gpt-4"];
-    if (gpt4Usage && gpt4Usage.cost > 100) {
-      opportunities.push({
-        id: "opt_cost_ai_downgrade",
-        type: "cost",
-        description: `High GPT-4 usage detected: $${gpt4Usage.cost.toFixed(2)} in 30 days`,
-        currentState: {
-          model: "gpt-4",
-          cost30Days: gpt4Usage.cost,
-          calls30Days: gpt4Usage.calls,
-        },
-        proposedChange: {
-          downgradeTo: "gpt-4o-mini",
-        },
-        expectedImpact: {
-          costSavings: gpt4Usage.cost * 0.9,
-          riskLevel: "low" as const,
-        },
-        recommendedAction: "human-review" as const,
-      });
-    }
-
-    // 2. Check for unused reconciliation jobs
-    const staleJobs = await prisma.reconJob.count({
-      where: {
-        status: "active",
-        lastRunAt: { lt: new Date(Date.now() - 90 * 24 * 60 * 60 * 1000) },
-      },
-    });
-
-    if (staleJobs > 10) {
-      opportunities.push({
-        id: "opt_cost_stale_jobs",
-        type: "cost",
-        description: `${staleJobs} stale reconciliation jobs detected (>90 days since last run)`,
-        currentState: {
-          staleJobCount: staleJobs,
-          threshold: "90 days",
-        },
-        proposedChange: {
-          action: "archive_or_delete",
-          targetCount: staleJobs,
-        },
-        expectedImpact: {
-          costSavings: staleJobs * 5, // $5 per job/month estimate
-          riskLevel: "low" as const,
-        },
-        recommendedAction: "human-review" as const,
-      });
-    }
-
-    // 3. Check for high-volume raw data
-    const unmappedCount = await prisma.rawRecord.count({
-      where: {
-        status: "failed",
-        createdAt: { gte: thirtyDaysAgo },
-      },
-    });
-
-    if (unmappedCount > 10000) {
-      opportunities.push({
-        id: "opt_cost_unmapped_cleanup",
-        type: "cost",
-        description: `High unmapped record count: ${unmappedCount} records in 30 days`,
-        currentState: {
-          unmappedRecords: unmappedCount,
-          storageCostEstimate: unmappedCount * 0.001,
-        },
-        proposedChange: {
-          action: "review_mappings",
-          autoCleanupThreshold: "30 days",
-        },
-        expectedImpact: {
-          costSavings: 25,
-          riskLevel: "low" as const,
-        },
-        recommendedAction: "human-review" as const,
-      });
-    }
-
-    return opportunities;
-  }
-
-  /**
-   * Find performance issues by analyzing job execution times and error rates
-   */
-  private async findPerformanceIssues(): Promise<OptimizationOpportunity[]> {
-    const opportunities: OptimizationOpportunity[] = [];
-
-    // 1. Check for error-prone connectors
-    const errorRates = await prisma.reconResult.groupBy({
-      by: ["connectorId"],
-      where: {
-        createdAt: { gte: new Date(Date.now() - 7 * 24 * 60 * 60 * 1000) },
-        status: { in: ["error", "failed"] },
-      },
-      _count: { id: true },
-    });
-
-    const totalRuns = await prisma.reconResult.groupBy({
-      by: ["connectorId"],
-      where: {
-        createdAt: { gte: new Date(Date.now() - 7 * 24 * 60 * 60 * 1000) },
-      },
-      _count: { id: true },
-    });
-
-    for (const errorStat of errorRates) {
-      const totalStat = totalRuns.find((t: any) => t.connectorId === errorStat.connectorId);
-      if (totalStat) {
-        const errorRate = errorStat._count.id / totalStat._count.id;
-        if (errorRate > 0.2) {
-          opportunities.push({
-            id: `opt_perf_errors_${errorStat.connectorId}`,
-            type: "performance",
-            description: `High error rate for connector ${errorStat.connectorId}: ${(errorRate * 100).toFixed(1)}%`,
-            currentState: {
-              connectorId: errorStat.connectorId,
-              errorRate,
-              errorCount7Days: errorStat._count.id,
-              totalCount7Days: totalStat._count.id,
-            },
-            proposedChange: {
-              action: "review_connector_config",
-              retryPolicy: "exponential_backoff",
-            },
-            expectedImpact: {
-              errorRateReduction: 0.5,
-              riskLevel: "low" as const,
-            },
-            recommendedAction: "human-review" as const,
-          });
-        }
-      }
-    }
-
-    // 2. Check for memory-intensive jobs
-    const largeJobs = await prisma.reconJob.findMany({
-      where: {
-        status: "completed",
-        executionTime: { gt: 300000 }, // > 5 minutes
-        createdAt: { gte: new Date(Date.now() - 7 * 24 * 60 * 60 * 1000) },
-      },
-      select: { id: true, name: true, executionTime: true },
-      take: 10,
-    });
-
-    for (const job of largeJobs) {
-      opportunities.push({
-        id: `opt_perf_memory_${job.id}`,
-        type: "performance",
-        description: `Memory-intensive job detected: ${job.name} (${job.executionTime}ms)`,
-        currentState: {
-          jobId: job.id,
-          jobName: job.name,
-          executionTimeMs: job.executionTime,
-        },
-        proposedChange: {
-          action: "enable_streaming",
-          batchSize: 1000,
-        },
-        expectedImpact: {
-          memoryReduction: 0.6,
-          riskLevel: "low" as const,
-        },
-        recommendedAction: "human-review" as const,
-      });
-    }
-
-    return opportunities;
-  }
-
-  /**
-   * Find capacity issues by analyzing queue depth and processing rates
-   */
-  private async findCapacityIssues(): Promise<OptimizationOpportunity[]> {
-    const opportunities: OptimizationOpportunity[] = [];
-
-    // 1. Check queue depth
-    const pendingJobs = await prisma.reconJob.count({
-      where: { status: "pending" },
-    });
-
-    const processingJobs = await prisma.reconJob.count({
-      where: { status: "processing" },
-    });
-
-    const queueRatio = pendingJobs / (processingJobs || 1);
-
-    if (queueRatio > 5) {
-      opportunities.push({
-        id: "opt_capacity_queue_depth",
-        type: "capacity",
-        description: `High queue depth: ${pendingJobs} pending, ${processingJobs} processing`,
-        currentState: {
-          pendingJobs,
-          processingJobs,
-          queueRatio,
-        },
-        proposedChange: {
-          action: "scale_workers",
-          targetWorkers: Math.ceil(pendingJobs / 10),
-        },
-        expectedImpact: {
-          throughputIncrease: 2.0,
-          riskLevel: "low" as const,
-        },
-        recommendedAction: "auto-apply" as const,
-      });
-    }
-
-    // 2. Check onboarding progress as a proxy for active users
-    const activeUsers24h = await prisma.onboardingProgress.count({
-      where: { updatedAt: { gte: new Date(Date.now() - 24 * 60 * 60 * 1000) } },
-    });
-
-    if (activeUsers24h > 100) {
-      opportunities.push({
-        id: "opt_capacity_users",
-        type: "capacity",
-        description: `High user load: ${activeUsers24h} active users in 24h`,
-        currentState: {
-          activeUsers24h,
-          threshold: 100,
-        },
-        proposedChange: {
-          action: "enable_cdn_caching",
-          cacheStaticAssets: true,
-        },
-        expectedImpact: {
-          loadReduction: 0.4,
-          riskLevel: "low" as const,
-        },
-        recommendedAction: "auto-apply" as const,
-      });
     }
 
     return opportunities;
